backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==0.8.11
PyPDF2==3.0.1
pydantic==2.5.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.0
google-cloud-language==2.12.0
spacy==3.7.2
uuid==1.30
requests==2.31.0